Fix RETRY - 1 issue (#118)
In some cases where retries are applied to an operation, the ending condition (used to log an error or throw an exception) is never reached because python ranges are exclusive of the ending bound. This PR fixes this issue and adds a -1 to the conditions so that the previously unreachable code is now used.