Implement patching in batches to improve performance. Done code changes for Aptitude package manager and did manual testing.
Code changes for Yum and Zipper is pending should be completed by next week.
Adding unit tests is pending.

Also, due to signature change of the method install_update_and_dependencies, there are some tests failing. I will work on those tests once we confirm the signature of the method.

Using Stopwatch implementation in this PR. The implementation of Stopwatch is in the PR https://github.com/Azure/LinuxPatchExtension/pull/161 which is going to be merged soon.