The changes contains following: (a) Incorporated PR comments
(b) Using all_package_versions instead of package_versions while finding version of the dependent package in the function include_dependencies in the file PatchInstaller.py. The list all_package_versions contains versions of all the packages available to install but the list package_versions contains only those packages which are selected by user. There might be dependency on the package which is not selected by user and hence all_package_versions should be used.

(c) Making a list containing not attempted packages and failed packages in batch patching process. In fallback to sequential patching, this list is used as the input packages to install. The method get_remaining_packages_to_install is no longer required. But due to this change, the newly available packages in brief time while batch patching was done are not handled. This was the case before this PR also, the newly added packages while installing the packages were not handled. To make it simple, it is better to not handle that case in this big PR. It is difficult to test this scenario.