Bugfix: Mitigate external Ubuntu Pro Client issues (#308)
**Background:** For VMs that are Ubuntu Pro client capable, there are 2
sets of issues that manifest as required security updates not getting
installed on machines_ (some machines).

1. Canonical is aware of packages normally seen as updates in the
default flow not being shown as required in pro client scans. There is
an explanation for this but the way we have taken a tight dependency on
pro client when it's functional causes these updates no to get
installed.

2. There are cases where wide swathes of security updates are not
getting detected by pro client. It was not clear if this was a pro
client issue in the past or an issue with our code. The additional code
that went in November helped identify that this was a pro client issue
when newer reports came in:
https://github.com/Azure/LinuxPatchExtension/pull/273

Both problems listed above are being resolved by not fully relying on
pro client and using a combined overlay of the default scanning
mechanism with whatever pro client reports. This is the 'best of both
worlds' approach. Extensive logging additions will help further reviews
with Canonical on pro client behaviors without affecting any customer
while a multi-stage resolution is ironed out.