Bugfix: Reboot Manager behavior - multiple bugfixes & error rate reduction (#310)
Fixing issues seen while testing:
[Bugfix: Mitigate external Ubuntu Pro Client
issues](https://github.com/Azure/LinuxPatchExtension/pull/308)

The main issue was that my test infra for that code failed to complete
patch installation with this error message:
_"Reboot failed to proceed on the machine in a timely manner."_

The machine then failed to accept ssh logins for 20+ minutes. I thought
the machine was dead, but it eventually came back suggesting a much
longer reboot time than expected. Tracing this further in production, it
suggested there were 500+ operations per week hitting the same issue and
getting reported as a terminal failure.

This prompted a closer look at all the Reboot Manager code, and PR
addresses every issue identified with it.

The changes in this PR:
1. Differentiating between all of the following:
(a) Reboot buffer in minutes = minimum time required to consider a
reboot. This was being overloaded incorrectly to also broadcast the time
delay to starting the reboot, which _could_ cause silent maintenance
window exceeds.
(b) Reboot notify timeout in minutes = introduced new to be deliberate
about the duration of the notification window.
(c) Reboot wait timeout in minutes (min & max) = time duration to wait
before considering an attempt to reboot a failure.

2. The effective reboot wait timeout is dynamic now - it's sits
somewhere between the min and max allowed and uses as much as the
remaining time in the maintenance window allows to allow for success.

3. Reboot manager code has been cleaned up to meet current coding
standards for the code base.

4. Reduce likelihood of timeout at the Compute RP by refreshing the
status as long as we're still waiting for a reboot and our process is
still running.

The configured values may be adjusted in the future to account for what
is seen at scale.