Bugfix: Mitigation for environments with Python Unbuffered I/O (#320)
A breaking change in the SUSE fork of the Azure Linux Agent caused
LinuxPatchExtension execution (and some others) to hang and fail when
executed against it. If a machine is running version 2.12.0.4 for either
the provisioning agent or the guest agent, it will be affected.

**Symptoms & Conditions:**
VMExtensionProvisioningErrors on SUSE 15, affecting all patch operations
on it too.
- Doesn’t happen if the agent version is 2.9.1.1 (both provisioning
agent, and guest agent)
- Doesn’t happen if it is 2.13.1.1 (both provisioning agent and guest
agent)
- Happens if PA is 2.12.0.4 (even if GA is 2.12.0.4 or higher, like the
latest 2.13.1.1)

This will happen 100% of the time on the latest image SUSE released
(April 17th) because it has 2.12.0.4 baked in and it auto-upgrades to
2.13.1.1 immediately (PA = 2.12.0.4, GA=2.13.1.1)

The immediate previous version (Jan 6th) starts with 2.9.1.1 and stays
on it unaffected. If upgraded manually directly to 2.13.1.1, everything
is good. If upgraded to 2.12.0.4, things break.

The scope of the problem is specifically long-running processes looking
like they are successfully starting (PID, etc.) but not actually doing
anything. (PA -> GA -> handler -> Core specifically)

**Customer Self-Mitigation Steps**
In the absence of this fix being available on their machines, customers
can self-mitigate by upgrading their agent (fully - both provisioning
agent and guest agent) to 2.13.1.1.

**What does the mitigation here entail?**
The extension handler was designed to capture stdout and stderr streams
in the event of an unforeseen failure in launching the core process.
This has been changed to default to not capturing those streams in the
first attempt (always succeeds on all machines today), and then _iff_ it
fails, resort to retrying with the streams to capture data for
troubleshooting.

**Are other extensions affected?**
Yes, others are. Discussions with SUSE are ongoing, so customers are
encouraged to mitigate the issue themselves if they hit provisioning
errors in other situations.